% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Enum}
\alias{Enum}
\title{Create Enum collections and lock bindings}
\usage{
Enum(l)
}
\arguments{
\item{l}{List of elements to add to \code{Enum} instance}
}
\description{
Create Enum collections and lock bindings
}
