% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializer.R
\name{getCommitMetaData}
\alias{getCommitMetaData}
\title{Get TASSEL standalone JAR metadata}
\usage{
getCommitMetaData(
 
    apiPath = "https://api.bitbucket.org/2.0/repositories/tasseladmin/tassel-5-standalone/src/master/",
  pageLen = 100
)
}
\arguments{
\item{apiPath}{Path to TASSEL Bitbucket API.}

\item{pageLen}{Number of pages to return. Defaults to \code{100}.}
}
\value{
A data.frame
}
\description{
Get commit history and download links
}
