% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getDirFields}
\alias{getDirFields}
\title{Get directory fields for installed package}
\usage{
getDirFields(pkg = "MGJars", sep = "/")
}
\arguments{
\item{pkg}{Package name}

\item{sep}{Directory separator character}
}
\description{
Get directory fields for installed package
}
